+++ hslua-module-path/src/HsLua/Module/Path.hs
@@ -68,6 +68,7 @@ fields =
 separator :: Field e
 separator = Field
   { fieldName = "separator"
+  , fieldType = "string"
   , fieldDescription = "The character that separates directories."
   , fieldPushValue = pushString [Path.pathSeparator]
   }
@@ -76,6 +77,7 @@ separator = Field
 search_path_separator :: Field e
 search_path_separator = Field
   { fieldName = "search_path_separator"
+  , fieldType = "string"
   , fieldDescription = "The character that is used to separate the entries in "
                     <> "the `PATH` environment variable."
   , fieldPushValue = pushString [Path.searchPathSeparator]